requirements.txt
# Client dependencies
PyQt6>=6.6.0
pillow>=10.0.0
pyautogui>=0.9.54
websockets>=11.0.0
requests>=2.31.0
qasync>=0.24.1

# Server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Shared dependencies
aiofiles>=23.2.1
python-dotenv>=1.0.0